+++ CHANGELOG.md
@@ -28,6 +28,8 @@
  - Changed name of method `rgb2hsl2` to `rgb2hsl`
  - Removed format parameter from `Grid` and `GridView` objects `encode` and `encodeSync` methods as it had no affect.
  - Added `active`, `queryable`, `clear_label_cache`, `minzoom`, and `maxzoom` property to `Layer` objects
+ - Added `compositeSync` to `VectorTile` object.
+ - Changed `composite` in `VectorTile` to accept a callback
 
 Notable changes in the Mapnik SDK include:
  - Changes: https://github.com/mapnik/mapnik/compare/30c6cf636c...7f597233c9